requirements.txt
google-api-python-client
google-cloud-bigquery
google-cloud-core
inflection
PyMySQL
python-dotenv
redis
setuptools
ua-parser
user-agents
wheel
apache-beam[gcp]